chore: rename no-multiple-objects-in-class
