+++ docs/rules/no-multiple-objects-in-class.md
@@ -1,17 +1,17 @@
 ---
 pageClass: rule-details
 sidebarDepth: 0
-title: vue/no-multiple-object-in-class
+title: vue/no-multiple-objects-in-class
 description: disallow to pass multiple objects into array to class
 ---
-# vue/no-multiple-object-in-class
+# vue/no-multiple-objects-in-class
 > disallow to pass multiple objects into array to class
 
 ## :book: Rule Details
 
 This rule disallows to pass multiple objects into array to class.  
 
-<eslint-code-block :rules="{'vue/no-multiple-object-in-class': ['error']}">
+<eslint-code-block :rules="{'vue/no-multiple-objects-in-class': ['error']}">
 
 ```vue
 <template>
@@ -33,5 +33,5 @@ Nothing.
 
 ## :mag: Implementation
 
-- [Rule source](https://github.com/vuejs/eslint-plugin-vue/blob/master/lib/rules/no-multiple-object-in-class.js)
-- [Test source](https://github.com/vuejs/eslint-plugin-vue/blob/master/tests/lib/rules/no-multiple-object-in-class.js)
+- [Rule source](https://github.com/vuejs/eslint-plugin-vue/blob/master/lib/rules/no-multiple-objects-in-class.js)
+- [Test source](https://github.com/vuejs/eslint-plugin-vue/blob/master/tests/lib/rules/no-multiple-objects-in-class.js)
