+++ lib/rules/no-multiple-objects-in-class.js
